# Speed up _mapping_repr