Creating a ordered list for filtering purposes using `.items()` turns out being rather slow. Use `.keys()` instead as that doesn't trigger a bunch of dataarray initializations.

- [x] Passes `pre-commit run --all-files`

Test case:

```python
import numpy as np
import xarray as xr

a = np.arange(0, 2000)
data_vars = dict()
for i in a:
    data_vars[f"long_variable_name_{i}"] = xr.DataArray(
        name=f"long_variable_name_{i}",
        data=np.arange(0, 20),
        dims=[f"long_coord_name_{i}_x"],
        coords={f"long_coord_name_{i}_x": np.arange(0, 20) * 2},
    )
ds0 = xr.Dataset(data_vars)
ds0.attrs = {f"attr_{k}": 2 for k in a}
```


Before:
```python
%timeit print(ds0)
14.6 s ± 215 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After:
```python
%timeit print(ds0)
120 ms ± 2.06 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```