@max-sixty 2021-08-02
Wow, that's quite the speed up! I haven't had time to understand why it's so dramatic, but let's merge unless there are any objections. Feel free to add a whatsnew @Illviljan