@dcherian 2021-08-02
> I haven't had time to understand why it's so dramatic

The old way creates all dataarrays (2000 in this case). This version creates `first_rows+last_rows` dataarrays. Very nice idea @Illviljan !