@dcherian 2021-08-02
I pushed a benchmark...

this branch:
[ 31.25%] ··· repr.Repr.time_repr 14.8±0.3ms
[ 37.50%] ··· repr.Repr.time_repr_html 164±1ms

main:
[ 31.25%] ··· repr.Repr.time_repr 41.6±0.5ms
[ 37.50%] ··· repr.Repr.time_repr_html 188±1ms