@Illviljan 2021-08-02
The benchmark uses (only) 100 arrays so I suppose there isn't as big of difference at that point. Did you get similar values as me with 2000, @dcherian ?