@dcherian 2021-08-02
3x is good! I reduced the size so it would run faster, we can't run a 15s benchmark every time we benchmark an older commit...

Can you add a whats-new note please?