@Illviljan 2021-08-02
Ok, if you got 15 seconds as well I'm satisfied. I was just worried I had done something else affecting my results.